#14001: Rename noc_params -> metal_noc_params
