+++ tests/tt_metal/tt_metal/perf_microbenchmark/dispatch/common.h
@@ -11,7 +11,7 @@
 #include "tt_metal/host_api.hpp"
 #include "tt_metal/impl/device/device.hpp"
 #include "tt_metal/impl/dispatch/cq_commands.hpp"
-#include "noc/noc_parameters.h"
+#include "noc/metal_noc_parameters.h"
 
 extern bool debug_g;
 extern bool use_coherent_data_g;
