+++ tt_metal/hw/firmware/src/wormhole/noc.c
@@ -6,7 +6,7 @@
 #include "noc.h"
 #include <stdint.h>
 #include <stdbool.h>
-#include "noc_parameters.h"
+#include "metal_noc_parameters.h"
 #include "risc_attribs.h"
