+++ tt_metal/impl/debug/watcher_device_reader.cpp
@@ -30,7 +30,7 @@
 #include "impl/device/device.hpp"
 #include "llrt/rtoptions.hpp"
 #include "noc/noc_overlay_parameters.h"
-#include "noc/noc_parameters.h"
+#include "noc/metal_noc_parameters.h"
 
 #include "watcher_device_reader.hpp"
