+++ tt_metal/llrt/tt_cluster.hpp
@@ -17,7 +17,7 @@
 #include "tt_metal/third_party/umd/device/tt_xy_pair.h"
 
 // clang-format off
-#include "noc/noc_parameters.h"
+#include "noc/metal_noc_parameters.h"
 #include "eth_interface.h"
 #include "eth_l1_address_map.h"
 #include "dev_msgs.h"
